storage/requirements.txt
llama-index==0.10.8
llama-index-llms-ollama==0.1.3
llama-index-embeddings-ollama==0.1.2
python-dotenv==1.0.1
gradio==4.19.2
tenacity==8.2.3 
